requirements.txt
mcp>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.12.0
redis>=5.0.0
openai>=1.3.0
httpx>=0.25.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pypdf2>=3.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pgvector>=0.2.0
numpy>=1.24.0
python-multipart>=0.0.6
celery>=5.3.0
kombu>=5.3.0
aiofiles>=23.2.0
tiktoken>=0.5.0
boto3>=1.28.0
kafka-python>=2.0.2
prometheus-client>=0.19.0
